requirements.txt
streamlit==1.28.1
langchain==0.0.350
requests==2.31.0
python-dotenv==1.0.0
langchain-community==0.0.5
